+++ pkg/plugin/plugin.go
@@ -4,11 +4,16 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"strings"
 
 	"github.com/argoproj-labs/rollouts-plugin-trafficrouter-glooedge/pkg/gloo"
 	"github.com/argoproj/argo-rollouts/pkg/apis/rollouts/v1alpha1"
 	pluginTypes "github.com/argoproj/argo-rollouts/utils/plugin/types"
 	"github.com/sirupsen/logrus"
+	gwv1 "github.com/solo-io/solo-apis/pkg/api/gateway.solo.io/v1"
+	v1 "github.com/solo-io/solo-apis/pkg/api/gloo.solo.io/v1"
+	"golang.org/x/exp/slices"
+	"google.golang.org/protobuf/types/known/wrapperspb"
 )
 
 const (
@@ -33,6 +38,17 @@ type DumbObjectSelector struct {
 	Labels    map[string]string `json:"labels" protobuf:"bytes,1,name=labels"`
 	Name      string            `json:"name" protobuf:"bytes,2,name=name"`
 	Namespace string            `json:"namespace" protobuf:"bytes,3,name=namespace"`
+}
+
+type destinationPair struct {
+	DestinationsParent *v1.MultiDestination
+	Canary             *v1.WeightedDestination
+	Stable             *v1.WeightedDestination
+}
+
+type routeTableWithDestinations struct {
+	RouteTable   *gwv1.RouteTable
+	Destinations []destinationPair
 }
 
 func (r *RpcPlugin) InitPlugin() pluginTypes.RpcError {
@@ -66,12 +82,22 @@ func (r *RpcPlugin) SetWeight(
 		}
 	}
 
-	if rollout.Spec.Strategy.Canary != nil {
-		err = r.handleCanary(ctx, rollout, desiredWeight, additionalDestinations, glooPluginConfig)
-		if err != nil {
-			return pluginTypes.RpcError{
-				ErrorString: fmt.Sprintf("failed canary rollout: %s", err),
-			}
+	if (glooPluginConfig.VirtualServiceSelector == nil && glooPluginConfig.RouteTableSelector == nil) ||
+		(glooPluginConfig.VirtualServiceSelector != nil && glooPluginConfig.RouteTableSelector != nil) {
+		return pluginTypes.RpcError{
+			ErrorString: "one of virtualService or routeTable selectors must be set in solo-io/glooedge plugin configuration",
+		}
+	}
+
+	if rollout.Spec.Strategy.Canary != nil && glooPluginConfig.VirtualServiceSelector != nil {
+		err = r.handleCanaryUsingVirtualService(ctx, rollout, desiredWeight, additionalDestinations, glooPluginConfig)
+	} else {
+		err = r.handleCanaryUsingRouteTables(ctx, rollout, desiredWeight, additionalDestinations, glooPluginConfig)
+	}
+
+	if err != nil {
+		return pluginTypes.RpcError{
+			ErrorString: fmt.Sprintf("failed canary rollout: %s", err),
 		}
 	}
 
@@ -116,4 +142,63 @@ func getStableServiceName(rollout *v1alpha1.Rollout) string {
 
 func getCanaryServiceName(rollout *v1alpha1.Rollout) string {
 	return rollout.Spec.Strategy.Canary.CanaryService
+}
+
+func (r *RpcPlugin) maybeCreateCanaryDestinations(
+	routeTables []routeTableWithDestinations, canaryName string) {
+
+	for i := range routeTables {
+		for j := range routeTables[i].Destinations {
+			if routeTables[i].Destinations[j].Canary != nil {
+				continue
+			}
+			routeTables[i].Destinations[j].Canary =
+				r.newCanaryDestination(routeTables[i].Destinations[j].Stable, canaryName)
+			routeTables[i].Destinations[j].DestinationsParent.Destinations =
+				append(routeTables[i].Destinations[j].DestinationsParent.GetDestinations(), routeTables[i].Destinations[j].Canary)
+		}
+	}
+}
+
+func (r *RpcPlugin) newCanaryDestination(stableDst *v1.WeightedDestination, canaryName string) *v1.WeightedDestination {
+	ret := stableDst.Clone().(*v1.WeightedDestination)
+	ret.GetDestination().GetUpstream().Name = canaryName
+	ret.Weight = &wrapperspb.UInt32Value{Value: uint32(0)}
+	return ret
+}
+
+func (r *RpcPlugin) getDestinationsInRoutes(
+	routes []*gwv1.Route,
+	rollout *v1alpha1.Rollout,
+	pluginConfig *GlooEdgeTrafficRouting) (ret []destinationPair) {
+
+	for _, r := range routes {
+		if len(pluginConfig.Routes) > 0 && !slices.Contains(pluginConfig.Routes, r.GetName()) {
+			continue
+		}
+
+		if r.GetRouteAction() == nil || r.GetRouteAction().GetMulti() == nil ||
+			r.GetRouteAction().GetMulti().GetDestinations() == nil {
+			continue
+		}
+
+		var stable, canary *v1.WeightedDestination
+		for _, dst := range r.GetRouteAction().GetMulti().GetDestinations() {
+			if dst.GetDestination() == nil || dst.GetDestination().GetUpstream() == nil ||
+				dst.GetDestination().GetUpstream().GetName() == "" {
+				continue
+			}
+			name := dst.GetDestination().GetUpstream().GetName()
+			if strings.EqualFold(getCanaryServiceName(rollout), name) {
+				canary = dst
+			} else if strings.EqualFold(getStableServiceName(rollout), name) {
+				stable = dst
+			}
+		}
+		if stable != nil {
+			ret = append(ret, destinationPair{DestinationsParent: r.GetRouteAction().GetMulti(), Stable: stable, Canary: canary})
+		}
+	}
+
+	return ret
 }
